serialize: change JSON field: urineTest -> urineTests
"urineTests" is the field name used in api/patients/:patientId, so this
should be changed to that for consistency. It wasn't obvious that the
Android app was running to this issue, because the app doesn't fail
when it can't find a urine test. However, an inspection of the database
on the app reveals that all the urine tests in the Android app are
NULL.