add lastEdited to Reading table for api/sync/readings
The lastEdited field is required to properly sync edits to existing
readings made by the Android app (i.e., when the Android app wants to
null out dateRecheckVitalsNeeded, that change should be picked up by
other Android clients that sync). This is because modifying
dateTimeTaken is problematic, since that does represent the time the
reading was taken.